Added test, mocks. added new operations changed some logic

+++ Calculator.BL/Logic/ICalculatorEngine.cs
@@ -3,6 +3,6 @@ namespace Calculator.BL.Logic
 {
     public interface ICalculatorEngine
     {
-        float Calculate(string[] chain);
+        double Calculate(string[] chain);
     }
 }
